Readd TODO
Still debating whether to do this or not, but leave it in as a reminder.

It seems like it would make everything less simple, with relatively
little benefit for me. But it's not unreasonable to expect this as an
extension of the original code.

For my use, I'd tend to use a likely unique address for some use and
just want the only message in there. Messages are deleted after some
time, so if you pick a unique enough address, it's unlikely there's
anything in there.